requirements.txt
docx2pdf==0.1.8
fastapi==0.115.4
httpx==0.27.0
Jinja2==3.1.4
opencv-python==4.10.0.84
openpyxl==3.1.5
pdfplumber==0.11.4
pydantic==2.8.2
PyMuPDF==1.24.10
pymupdf4llm==0.0.17
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
starlette==0.41.2
tqdm==4.66.4
uvicorn==0.32.0
python-multipart=0.0.17
